Changed Refactored Algorithms. Path changed from Array<Node> to Queue<Node> becuse it makes more sense. All Algorithms build path using a array contining the nodes parent which was done in a* and dijkstras. Moved each algorithm in to a seperate file to make it more readable.
