+++ src/path.ts
@@ -1,5 +1,6 @@
 import { Node } from "./graphs";
 import { Pair } from "./list";
+import { Queue } from "./queue_array";
 
 /**
  * A Path is a Array containing the "Path" to a given node.
@@ -9,7 +10,7 @@ import { Pair } from "./list";
  * 
  * Path can be Empty ([]) if no Path to the node was found.
  */
-export type Path = Array<Node>;
+export type Path = Queue<Node>;
 
 
 /**
