+++ tests/graphs.test.ts
@@ -1,13 +1,11 @@
 import {
     undirected, EdgeList,
     MatrixGraph, ListGraph,
-    mg_from_edges, mg_new,
-    lg_from_edges, lg_new, lg_transpose,
-    lg_depth_first, lg_breadth_first, lg_dijkstra_path, lg_a_star_path
+    mg_from_edges, mg_new, Node,
+    lg_from_edges, lg_new, lg_transpose, create_path
 } from "../src/graphs";
-import { head, is_null, list, tail } from "../src/list";
-import { Maze, maze_to_listgraph, new_maze } from "../src/maze";
-import { Result, Stream } from "../src/path";
+import { list } from "../src/list";
+import { queue } from "../src/queue_array";
 
 
 /**
@@ -452,923 +450,29 @@ test("test transposing a list graph", () => {
     );
 });
 
-
-const a = 0;
-const b = 1;
-const c = 2;
-const d = 3;
-const e = 4;
-const f = 5;
-const g = 6;
-const h = 7;
-const listgraph: ListGraph = {
-    adj: [
-        list(b, d), // A: 0
-        list(c, f), // B: 1
-        list(g), // C: 2
-        list(c, e, g), // D: 3
-        list(g), // E: 4
-        list(e, h), // F: 5
-        list(), // G: 6
-        list(g) // H: 7
-    ],
-    size: 8
-};
-
-function eval_stream(stream: Stream<Result>): Result {
-    let return_value = stream;
-    while (!head(return_value).is_done) {
-        let stream_tail = tail(return_value);
-        if (!is_null(stream_tail)) {
-            return_value = stream_tail();
-        } else {
-            return head(return_value);
-        }
-    }
-    return head(return_value);
-}
-
-test("test listgraph breadth first search", () => {
-    let result: Result = eval_stream(
-        lg_breadth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(a)
-                    ],
-                size: 2
-            } as ListGraph,
-            a,
-            b
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b]
-    );
-
-    result = eval_stream(
-        lg_breadth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c),
-                        list(a)
-                    ],
-                size: 3
-            } as ListGraph,
-            a,
-            c
-        )
-    );
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c]
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c),
-                        list(a)
-                    ],
-                size: 3
-            } as ListGraph,
-            b,
-            c
-        )
-    );
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [b, c]
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c),
-                        list()
-                    ],
-                size: 3
-            } as ListGraph,
-            c,
-            b
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        Array<Node>()
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, a),
-                        list(a, b)
-                    ],
-                size: 3
-            } as ListGraph,
-            c,
-            b
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [c, b]
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, d),
-                        list(e),
-                        list(e),
-                        list()
-                    ],
-                size: 5
-            } as ListGraph,
-            a,
-            e
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c, e]
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, d),
-                        list(e, f),
-                        list(e),
-                        list(f),
-                        list()
-                    ],
-                size: 6
-            } as ListGraph,
-            a,
-            f
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c, f]
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, d),
-                        list(e),
-                        list(e, f),
-                        list(f),
-                        list()
-                    ],
-                size: 6
-            } as ListGraph,
-            a,
-            f
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, d, f]
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(listgraph, a, g)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, d, g]
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(listgraph, a, h)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, f, h]
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(listgraph, a, c)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c]
-    );
-
-
-    result = eval_stream(
-        lg_breadth_first(listgraph, a, e)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, d, e]
-    );
-});
-
-
-test("test listgraph depth first search", () => {
-    let result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(a)
-                    ],
-                size: 2
-            } as ListGraph,
-            a,
-            b
-        )
-    );
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c),
-                        list(a)
-                    ],
-                size: 3
-            } as ListGraph,
-            a,
-            c
-        )
-    );
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c),
-                        list(a)
-                    ],
-                size: 3
-            } as ListGraph,
-            b,
-            c
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [b, c]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c),
-                        list()
-                    ],
-                size: 3
-            } as ListGraph,
-            c,
-            b
-        )
-    );
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        Array<Node>()
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, a),
-                        list(a, b)
-                    ],
-                size: 3
-            } as ListGraph,
-            c,
-            b
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [c, b]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, d),
-                        list(e),
-                        list(e),
-                        list()
-                    ],
-                size: 5
-            } as ListGraph,
-            a,
-            e
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, d, e]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, d),
-                        list(e, f),
-                        list(e),
-                        list(f),
-                        list()
-                    ],
-                size: 6
-            } as ListGraph,
-            a,
-            f
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, d, e, f]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, d),
-                        list(e),
-                        list(e, f),
-                        list(f),
-                        list()
-                    ],
-                size: 6
-            } as ListGraph,
-            a,
-            f
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, d, f]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(b, f),
-                        list(c),
-                        list(b),
-                        list(c),
-                        list(d),
-                        list(e)
-                    ],
-                size: 6
-            } as ListGraph,
-            a,
-            c
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, f, e, d, c]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(
-            {
-                adj:
-                    [
-                        list(f, b),
-                        list(c),
-                        list(b),
-                        list(c),
-                        list(d),
-                        list(e)
-                    ],
-                size: 6
-            } as ListGraph,
-            a,
-            c
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(listgraph, a, g)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, d, g]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(listgraph, a, h)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, f, h]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(listgraph, a, c)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, d, c]
-    );
-
-
-    result = eval_stream(
-        lg_depth_first(listgraph, a, e)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, d, e]
-    );
-});
-
-// Same path as BFS
-test("test listgraph dijkstra_path search", () => {
-    let result: Result = eval_stream(
-        lg_dijkstra_path(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(a)
-                    ],
-                size: 2
-            } as ListGraph,
-            a,
-            b
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b]
-    );
-
-    result = eval_stream(
-        lg_dijkstra_path(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c),
-                        list(a)
-                    ],
-                size: 3
-            } as ListGraph,
-            a,
-            c
-        )
-    );
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c]
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c),
-                        list(a)
-                    ],
-                size: 3
-            } as ListGraph,
-            b,
-            c
-        )
-    );
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [b, c]
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c),
-                        list()
-                    ],
-                size: 3
-            } as ListGraph,
-            c,
-            b
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        Array<Node>()
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, a),
-                        list(a, b)
-                    ],
-                size: 3
-            } as ListGraph,
-            c,
-            b
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [c, b]
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, d),
-                        list(e),
-                        list(e),
-                        list()
-                    ],
-                size: 5
-            } as ListGraph,
-            a,
-            e
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c, e]
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, d),
-                        list(e, f),
-                        list(e),
-                        list(f),
-                        list()
-                    ],
-                size: 6
-            } as ListGraph,
-            a,
-            f
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c, f]
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(
-            {
-                adj:
-                    [
-                        list(b),
-                        list(c, d),
-                        list(e),
-                        list(e, f),
-                        list(f),
-                        list()
-                    ],
-                size: 6
-            } as ListGraph,
-            a,
-            f
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, d, f]
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(listgraph, a, g)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, d, g]
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(listgraph, a, h)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, f, h]
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(listgraph, a, c)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, b, c]
-    );
-
-
-    result = eval_stream(
-        lg_dijkstra_path(listgraph, a, e)
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [a, d, e]
-    );
-
-
-    let maze = new_maze(3);
-    maze.matrix[0][1] = false;
-    maze.matrix[1][1] = false;
-    maze.matrix[2][1] = false;
-    result = eval_stream(
-        lg_dijkstra_path(
-            maze_to_listgraph(maze),
-            0,
-            8
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        []
-    );
-});
-
-
-test("test listgraph A* search", () => {
-    let maze: Maze;
-    maze = new_maze(2);
-    let result: Result = eval_stream(
-        lg_a_star_path(
-            maze_to_listgraph(maze),
-            maze,
-            0,
-            3
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [0, 1, 3]
-    );
-
-
-    maze = new_maze(3);
-    result = eval_stream(
-        lg_a_star_path(
-            maze_to_listgraph(maze),
-            maze,
-            0,
-            8
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [0, 1, 2, 5, 8]
-    );
-
-
-    maze = new_maze(3);
-    maze.matrix[0][1] = false;
-    result = eval_stream(
-        lg_a_star_path(
-            maze_to_listgraph(maze),
-            maze,
-            0,
-            8
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [0, 3, 4, 5, 8]
-    );
-
-
-    maze = new_maze(3);
-    maze.matrix[0][1] = false;
-    maze.matrix[1][1] = false;
-    result = eval_stream(
-        lg_a_star_path(
-            maze_to_listgraph(maze),
-            maze,
-            0,
-            8
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [0, 3, 6, 7, 8]
-    );
-
-
-    maze = new_maze(3);
-    maze.matrix[0][1] = false;
-    maze.matrix[1][1] = false;
-    result = eval_stream(
-        lg_a_star_path(
-            maze_to_listgraph(maze),
-            maze,
-            0,
-            2
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        [0, 3, 6, 7, 8, 5, 2]
-    );
-
-
-    maze = new_maze(3);
-    maze.matrix[0][1] = false;
-    maze.matrix[1][1] = false;
-    maze.matrix[2][1] = false;
-    result = eval_stream(
-        lg_a_star_path(
-            maze_to_listgraph(maze),
-            maze,
-            0,
-            8
-        )
-    );
-
-    expect(
-        result.path_so_far
-    ).toStrictEqual(
-        []
-    );
+test("Creating a path given a previous_nodes array", () => {
+    let prev = Array<Node>(1);
+    expect(create_path(0, 0, prev)).toStrictEqual([0, 1, [0]]);
+    expect(create_path(0, 0, prev)).toStrictEqual(queue(0));
+
+
+    prev = Array<Node>(2);
+    prev[1] = 0;
+    expect(create_path(0, 1, prev)).toStrictEqual([0, 2, [0, 1]]);
+    expect(create_path(0, 1, prev)).toStrictEqual(queue(0, 1));
+
+
+    prev = Array<Node>(3);
+    prev[1] = 0;
+    prev[2] = 1;
+    expect(create_path(0, 2, prev)).toStrictEqual([0, 3, [0, 1, 2]]);
+    expect(create_path(0, 2, prev)).toStrictEqual(queue(0, 1, 2));
+
+
+    prev = Array<Node>(3);
+    prev[1] = 0;
+    prev[2] = 1;
+    prev[3] = 2;
+    expect(create_path(0, 3, prev)).toStrictEqual([0, 4, [0, 1, 2, 3]]);
+    expect(create_path(0, 3, prev)).toStrictEqual(queue(0, 1, 2, 3));
 });
